fix: apply default streaming value consistently in sendMessage execution. TASK: AYC-20
