--- ayunis-core-backend/src/domain/storage/application/use-cases/upload-file/upload-file.command.ts
@@ -1,9 +1,0 @@
-import { ScopeType } from '../../../domain/value-objects/scope-type.enum';
-
-export class UploadFileCommand {
-  constructor(
-    public readonly file: Express.Multer.File,
-    public readonly scopeType: ScopeType,
-    public readonly scopeId?: string,
-  ) {}
-}
