--- ayunis-core-backend/src/domain/storage/domain/value-objects/file-type-category.enum.ts
@@ -1,5 +1,0 @@
-export enum FileTypeCategory {
-  IMAGE = 'image',
-  DOCUMENT = 'document',
-  DATA = 'data',
-}
