--- ayunis-core-backend/src/domain/storage/domain/value-objects/scope-type.enum.ts
@@ -1,7 +1,0 @@
-export enum ScopeType {
-  THREAD = 'thread',
-  // AGENT = 'agent',
-  // SOURCE = 'source',
-  // USER = 'user',
-  // Add more scope types as needed
-}
